Small Reorganisation of code
